fix for multicolumn save/load
We now save the columns field as the maximum number of columns for which
a grid layout has been specified.

If, when loading a saved freeboard, the specified number of columns in the
columns field is greater than the available width, we default to the maximum
number of columns that fit on the current screen.