Fix multi-column resizing issue on zoomed browsers.
On some browsers (Chrome) the calculated width of DOM elements can be off by
a pixel or two when the browser is zoomed far out. Likely due to a floating
point error in the pixel scaling. This manifests as a broken add column button:
we set the width to be exactly wide enough to accommodate the columns we want,
but when gridster checks the container width it finds it to be a pixel too
small. Add a few pixels buffer to the max-width we set on the gridster
container to workaround this.